requirements.txt
fastapi
uvicorn
gunicorn
mysql-connector-python
requests
pandas
